Improve the algorithm to detect font strings
Don't check for bold or italic, consider it the style name. This makes it
a bit more difficult to be able to find what's style name and what's font
family.
Loop words checking where the font family ends and assume the stylename is
the rest until the font size

PS: Why wouldn't they split the fields using ','? :/