Save gtk3 configuration to dconf as well
Summary:
Currently all gtk3 configuration is stored to a config file in ini format, while
this configuration is still used and supported, it is not used by applications
running under flatpak. To have this configuration available to sandboxed applications
we have to use dconf, which is exposed to flatpak applications.

BUG: 387417

Reviewers: apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9163