Sort icon and cursor themes case-insensitively
Summary:
Stick a QSortFilterProxyModel between the model and the view, and tell it to sort case-insensitively. Thanks to https://stackoverflow.com/a/5732026/2934226 for providing the solution.

BUG: 368600
FIXED-IN: 5.14.0

Test Plan: Could not test as I was unable to locate a working GTK 2 or 3 theme that started with a lowercase letter. But theoretically this code should do the trick, and it does not appear to cause any regressions at least.

Reviewers: apol, #plasma, broulik

Reviewed By: apol

Subscribers: wbauer, cfeck, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15598