Sync dark theme preference for GTK3 applications
Summary:
Previously preference of dark GTK3 theme was set by checkbox in GTK KCM.  Now, the checkbox is removed and preference is set according to color scheme, the user set in Colors KCMm, using the same heuristic, used for filter combobox in that KCM.

This makes sense, because, suppose the user set Adwaita GTK3 theme, and their color scheme is something dark. Therefore their overall theme is supposedly dark, so Adwaita needs to be dark too.

Note, however, that this preference is somewhat buggy. It does not work in flatpak apps, it can't be autoreloaded with xsettingsd, and also it is applied after some time has passed (this is hypothetical). You can check these statements, using the present checkbox in GTK KCM.

Depends on D25872

Test Plan:
{F7816371}
# Apply Adwaita GTK3 Theme
# Reload kded5
# Apply dark color scheme
# Wait some time
# (Re)open GTK3 app
# App should use Dark color scheme
# Apply light color scheme
# Wait some time
# Reopen GTK3 app
# App should use Light color scheme

Reviewers: apol, cblack, broulik, #vdg

Reviewed By: apol, cblack

Subscribers: ngraham, GB_2, plasma-devel

Tags: #plasma

Maniphest Tasks: T10611

Differential Revision: https://phabricator.kde.org/D25945