Upgrade theme to Breeze GTK on startup
Also apply it, if the config is empty. This is a port of kconf_update
script from `breeze-gtk` repo, but this time it use gtkconfig kded
module methods to manipulate config (to avoid code duplication and to
write config to all possible configuration storages).

Also kded module updates config for theme on every start, if it is
empty.

Summary:
BUG: 416635
FIXED-IN: 5.18.0

Test Plan:

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26891