Use only GTK Module for managing window decorations CSS
Now we are no longer use gtk.css file inside ~/.config/gtk-3.0/ directory
for window decorations. To load decorations into GTK apps we now exclusively
use GTK module.

The module is a GTK-app addon, that loads the buttons on application startup,
then it reacts to changes of windows_decorations.css file in config directory
and loads and unloads it dynamically depending on its presence.

For example: file is present - it is loaded on the startup, file is deleted -
it is unloaded, file is created or changed - it is loaded, while the past
version of it is unloaded.

File presence is controlled by kde-gtk-config kded5 module.

BUG: 428322
BUG: 428120
BUG: 428842
(cherry picked from commit 604ff30f03769c191cb51f113157073c096942f4)